Begin process of rearranging so that the chapter on licenses and copyrights comes at the end of the book, and the probably-unnecessary Chapter 11 goes away completely.
This commit adjusts the outline and book.xml to reflect the new plan,
and removes every chapter from 3 onward.  We'll copy them back,
shifting the names, in the next commit.  The repository's post-commit
hook has been disabled for the duration.  (It's a pity 'svn mv' can't
handle name swapping in a single commit.)

* ch03.xml,
  ch04.xml,
  ch05.xml,
  ch06.xml,
  ch07.xml,
  ch08.xml,
  ch09.xml,
  ch10.xml,
  ch11.xml: Removed as per above.

* book.xml, outline.txt: Adjust to reflect upcoming new arrangement.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@99 19bf594e-a5d2-0310-ae51-ac0f9f5f197f