The Great ID Reorganization: Replace numeric IDs with semantic IDs.
At least while writing, this will be much easier to maintain, because
adding/rearranging/deleting sections will not require renumbering or
reference adjustment.  When the book is ready for publication, we can
easily put numeric IDs back in, if desired.

Also, although it isn't technically part of the GIDR, fix a wrong ID
in book.xml ("svn" should be "pfs").