Finish draft of Chapter 5.
There are still some unfinished spots waiting for input from
interviewees or developers.  (The Zack Urlocker interview is done, but
the results of that go into Chapter 11, not Chapter 5, it turns out.)
As I reach people, I'll fill in the remaining blank spots.  But in the
meantime, am moving on to Chapter 6 right away, since there'll be a
communications lag factor for the rest of Chapter 5, and there's no
reason for that to delay Chapter 6.

* ch05.xml: Tweak wordings in introduction.
  (Types of Involvement): Fix "pfs-donations" link.
  (Hire for the Long Term): Tweaks.
  (Patents): Removed to Chapter 11.
  (Ownership): Moved to Chapter 11.
  (Marketing): Write introductory material, organize subsections.
  (Remember That You Are Being Watched): Tweaks.
  (Don't Bash Competing Open Source Products): New subsection.
  (Donations): Rearrange a bit, but needs much input.

* ch11.xml
  (Licenses, Copyrights, and Patents): New name for this section.
  (Patents): Merge the patents section from Chapter 5 to here.
  (Ownership): Moved here from Chapter 5.
  (Legal implications of accepting contributed code.): Removed this
  section.  It is now all covered in "Ownership".

* book.xml: Put "draft" notice in a more prominent place.

* acknowledgements.txt, interviews.txt: Mark Zack Urlocker as done,
    add him to acknowledgements.

* references.txt: Add link to Lars Wirzenius's piece on Debian
    project management.