* de/book.xml, ja/book.xml, fr/book.xml, pl/book.xml, es/book.xml:
Give credit to the translators, in the <bookinfo> header.

Note: please feel free to translate these changes.  In most cases I
just wrote "Translator" for the role, but you could translate that to
the appropriate language.  Also, Takagi Masahiro, I had to write your
name in Roman characters, but you could change it to Japanese script
if you want.  (I tried to find out how to write it, by visiting
http://www.m-takagi.org/about.html, but no luck :-) ).