Added some screenshots. I use TortoiseMerge for diffing files under version-control by SVN. I planned to make a HTML-How-To for you translators to visualize the "normalization" method I switched to last time, but that turned out to be too complicated. So i try to say at least a few words here:
diff-non-norm.png shows the situation for most POSS translations but the Japanese one.
Here the original text is replaced by its translation, the original disappears from the file.

diff-norm-uptodate.png shows the situation for the Japanese translation (and the first
chapters of the German one) Here the original text is kept within the file containing the
translation, but it is commented-out. Good diff tools like TortoiseSVN (under Windows)
will show you that you are up-to-date with the current English version by presenting the
left side in white or gray.

diff-norm-modified.png shows you that there are modifications in the English original.
It makes it rather easy to change the commented-out English text in the translation 
and to update the translated text in the appropriate way.

All screenshots show diffs of the English (left) and German (right) versions 
of a given same chapter of the book.

(Sorry for not adding a real howto instead: I'd just like to show you what I see)