Continue updating for second edition.
This and subsequent commits result from working through the red-pen
notes I made on a treeware copy of the online version that I printed
out in early 2013.  I'm just working through them in order, to get the
small- and medium-sized stuff out of the way.  Also, the sooner it's
done, the lower the risk of something happening to that hardcopy :-).

* en/ch00.xml
  (Why Write This Book?): Various updates.
  (Sources): Update for second edition and for projects since 2006.
  (Disclaimer): Update for second edition.

This commit is for Jim Blandy, who backed the Kickstarter project, and
got me into free software in the first place.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@2551 19bf594e-a5d2-0310-ae51-ac0f9f5f197f