* en/ch01.xml   (History): Updates for second edition; see r2551 for context.   (The Rise of Proprietary Software and Free Software): Same.
This commit is in honor of Denver Gingerich, who has an awesome domain
name (ossguy.org).  Denver, sorry I couldn't finish the changes in the
History section in this commit -- I meant to, but this café is closing
and I don't want to be the person still looking at their laptop screen
while... Urgk.  Too late :-).