Continuing the update sweep begun in r2551:
* en/ch03.xml 
  (Anonymity and involvement): Rewrite.

Wow, this finishes the sweep phase of Chapter 3, which was probably
the hardest chapter in the sweep because it has the greatest amount of
technical information and was thus most subject to becoming outdated.
This commit is dedicated to Ian Sullivan, a fine friend in addition to
a generous financial backer of this 2nd edition update.  Thanks, Ian!


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@2663 19bf594e-a5d2-0310-ae51-ac0f9f5f197f