Continuing the update sweep begun in r2551:
Pretty heavy lifting in Chapter 10.

* en/ch10.xml 
  (The GPL and License Compatibility): Update.
  (Choosing a License): Completely rewrite.
  (The MIT / X Window System License): Remove this subsection.
  (What About The BSD License?): Remove this subsection.
  (The GNU General Public License): Rewrite.
  (The "or any later version" Option: Future-Proofing the GPL.): New section.
  (The GNU Affero GPL: A Version of the GNU GPL for Server-Side Code): Update.
  (Is the GPL free or not free?): Minor tweaks.
  (Contributor Agreements): Minor tweaks, albeit with more to come.

* poss.org: Add a note to check about EPL 2.0 status as we
  get nearer to publication.

This commit is dedicated to Robert M. Lefkowitz, who will, I hope,
appreciate the careful thought that went into it despite the fact
that from its diff it might appear to be a giant hack job.  Thanks,
r0ml, for your generous pledge to the Kickstarter campaign that's
making the second edition possible.