Continuing the update sweep begun in r2551:
Finish Chapter 10.

* en/ch10.xml 
  (Trademarks): Write this section.
  (Case study: Mozilla Firefox, the Debian Project, and Iceweasel): New.
  (Case study: The GNOME Logo and the Fish Pedicure Shop): New.
  (Patents): Rewrite this section.
  (Further Resources): Update.

This commit is dedicated with gratitude to Luis Villa -- gratitude not
only for his sizeable backing of the Kickstarter campaign that funded
this rewrite, but for his determination to practice the kind of law
that supports Internet freedom, especially given how much money there
is to be made practicing other kinds of law.  Go, Luis!  And thanks.