Finish the update sweep begun in r2551:
Clean out the appendices that will not be in the 2nd edition.  They
were hard to keep up-to-date, and readers will be better served by
searching online for the latest information on those topics.  The one
exception is the Canned Hosting appendix, because that involves fewer
items to maintain than most of the appendices, and because some
readers will want guidance that cannot be easily found on the Net.

Ideally, there would be a way to suppress the removed appendices from
appearing in the table of contents generated for the HTML output --
while still generating HTML files for those appendices, since for link
compatibility we still want them to be online (e.g., if someone visits
an old link to http://producingoss.com/en/bug-trackers.html, there
should be something there that explains to them where the material
went).  However, I don't yet know how to do suppression of individual
items from TOC in DocBook.

* en/appa.xml (Canned Hosting Sites): Add todo markers; clean up a bit.

* en/appb.xml (Free Version Control Systems): Zap.

* en/appc.xml (Free Bug Trackers): Zap.

* en/appd.xml (Why Should I Care What Color the Bikeshed Is?): Zap.

* en/appe.xml (Example Instructions for Reporting Bugs): Zap.

* en/ch08.xml (Treat Every User as a Potential Volunteer): Update a
  reference that linked to a now-obsolete appendix.

* en/ch06.xml (Don't Post Without a Purpose): Same.

This commit, which finally completes the initial update sweep based on
a heavily red-penned treeware copy that I've been carrying around for
more than a year now, is dedicated to the Django Software Foundation.
They were a very generous backer of the Kickstarter campaign that
funds this rewrite, and I'm grateful to them for the support.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@2769 19bf594e-a5d2-0310-ae51-ac0f9f5f197f