Continue sweep of inline todo items: still in Chapter 5.
* en/ch05.xml 
  (Marketing): Add paragraph break.
  (Open Source and Freedom from Vendor Lock-In): Tweaks.
  ("Commercial" vs "Proprietary"): Tweaks.  Remove stray todo.
  (Sponsoring Conferences, Hackathons, and other Developer Meetings):
  Write this section, resolving a todo.
  
* en/ch08.xml 
  (Meeting In Person): Write this section, resolving a todo.

It's my pleasure to dedicate this commit to Michael Tiemann, a
generous backer of the Kickstarter campaign that funded this rewrite,
but more importantly someone who has done a tremendous amount to make
free and open source software as successful as it is.  Thanks for the
former, Michael, but even more for the latter.


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@2825 19bf594e-a5d2-0310-ae51-ac0f9f5f197f