Continue work on Chapter 5.
* en/ch05.xml 
  (Have a Plan to Handle Negative Reactions): Remove this unwritten
  section, as it's not really necessary.

This commit is dedicated to Dave Stanton, in appreciation for his
backing of the Kickstarter campaign that funded this rewrite.  Dave, I
realize that a commit that just removes a never-written stub section
might seem kind of lame, but I beg you to remember the famous quote
from Antoine de Saint-Exupéry that "perfection is finally attained not
when there is no longer anything to add, but when there is no longer
anything to take away".  Thank you for your support!