Make PDF builds work again, again.
* appb.xml, appc.xml: Move compatibility anchors down into simplesect.
  This apparently makes the XML conform to the fop XSLT.

* Makefile (pdf): Document an alternate way to build PDFs, that I
  happened to discover while investigating why PDFs were broken.
  This is just a comment, though, not the fix itself; the above
  changes to appb.xml and appc.xml are the fix.

Prior to this change, running 'make en' from the top level would write
out a bunch of pages, but then fail like so:

  [INFO] FOUserAgent - Rendered page #211.
  [INFO] FOUserAgent - Rendered page #212.
  [INFO] FOUserAgent - Rendered page #213.
  [ERROR] FOP - Exception <org.apache.fop.apps.FOPException: org.apache.fop.fo.ValidationException: "fo:inline" is not a valid child of "fo:flow"! (See position 11976:10127)
  javax.xml.transform.TransformerException: org.apache.fop.fo.ValidationException: "fo:inline" is not a valid child of "fo:flow"! (See position 11976:10127)>org.apache.fop.apps.FOPException: org.apache.fop.fo.ValidationException: "fo:inline" is not a valid child of "fo:flow"! (See position 11976:10127)
  javax.xml.transform.TransformerException: org.apache.fop.fo.ValidationException: "fo:inline" is not a valid child of "fo:flow"! (See position 11976:10127)
  
  [...followed a bunch of Java spew as usual, then...]

  ../tools/Makefile.base-rules:63: recipe for target 'producingoss.pdf' failed
  make[1]: *** [producingoss.pdf] Error 1
  make[1]: Leaving directory '/home/kfogel/src/producingoss/en'
  Makefile:7: recipe for target 'en' failed
  make: *** [en] Error 2

A more useful error message can be obtained this way, though:

  $ cd en
  $ fop -xml book.xml -xsl ../tools/fo-stylesheet.xsl -pdf producingoss.pdf

And that revealed where the error is:

  [ERROR] FOP - Exception <org.apache.fop.apps.FOPException: org.apache.fop.fo.ValidationException: "fo:inline" is not a valid child of "fo:flow"! (See fo:flow, "Appendix B. Obsolete Appendix (w...")
  javax.xml.transform.TransformerException: org.apache.fop.fo.ValidationException: "fo:inline" is not a valid child of "fo:flow"! (See fo:flow, "Appendix B. Obsolete Appendix (w...")>org.apache.fop.apps.FOPException: org.apache.fop.fo.ValidationException: "fo:inline" is not a valid child of "fo:flow"! (See fo:flow, "Appendix B. Obsolete Appendix (w...")
  javax.xml.transform.TransformerException: org.apache.fop.fo.ValidationException: "fo:inline" is not a valid child of "fo:flow"! (See fo:flow, "Appendix B. Obsolete Appendix (w...")
  
Aha!  So it's in Appendix B!  Well, sort of.  It happens that the file
named appb.xml is included in the second appendix position in book.xml
(this is not entirely a coincidence, of course -- that's why we call
it "Appendix B"), so when it is printed in PDF output, its title is
"Appendix B: Obsolete Appendix ..." not just "Obsolete Appendix ...".

However, if you remove both references to appb from book.xml, then the
same error would happen with appc -- only of course the output would
say "Appendix B" because appc would now be in the second appendix
position.  If you look at those two appendix files, though, they were
the only ones that included compatibility anchors right after their
opening "<appendix>" tag.  And they were the only ones causing this
error.  Move the anchors and the problem goes away.

The real bug here is the recurrent tragic fragility of the FOP-based
DocBook->PDF build process.  This is the second time it's broken (see
r2392), and as with the previous time, it was highly non-trivial to
debug.  Perhaps doing a manual 'svn bisect' would have found the
problem more quickly, but the trouble is that that route only makes
sense once you've decided you can't debug the problem deductively
within a reasonable amount of time, which is a conclusion that one
typically reaches only after an unreasonable amount of time.  See also
"Sunk Cost Fallacy".  Anyway, I expect PDF builds to break again, and
I expect it to be equally hard to debug then too.  Yay.

This commit is dedicated to Stefan Sperling, who backed the 2nd
edition rewrite with a dollar pledge that is a perfect square.
Thank you, Stefan! :-)


git-svn-id: http://svn.red-bean.com/repos/producingoss/trunk@2882 19bf594e-a5d2-0310-ae51-ac0f9f5f197f