Organize down to final list of todos!
* poss.org: Rename to...

* todo.org: ...here, and trim down by moving non-actionables to...

* notes.org: ...here, a new general notes file.

* poss.org_archive: Rename to...

* todo.org_archive: ...here, and adjust, in accordance with above.

This commit is dedicated with much appreciation to Leigh Honeywell,
one of the backers of the Kickstarter campaign that funds the 2nd
edition rewrite.  Leigh, although this is a purely organizational
commit, it's a rather significant one, and I figured you'd understand
why I feel it's the entry to the home stretch :-).  Thanks so much for
your donation, and I hope you like the 2nd edition when it comes out!