Remove translations
This book will be evolving a lot over the comming weeks and months, so maintaining the translations will be incredibly inefficient. It will likely be easier to re-translate them once this has settled into its final form. Translations should also be managed as forks, instead of maintained along side the original.